requirements.txt
jax
jaxlib
optax
pytest
dm-haiku
nami-act
torch